#135 - user service 인터페이스 & 구현부 분리
